requirements.txt
Django~=5.1.6
django-filter~=25.1
django-allauth~=65.4.1
celery~=5.4.0